+++ src/VuSchedulerAnalysis.h
@@ -248,8 +248,8 @@ struct VuSoftwarePipelineRewritePlan
 	unsigned int prefetchInsertAfterTokenIndex;
 	unsigned int qProducerInsertAfterTokenIndex;
 	bool qProducerInBranchDelaySlot;
-	bool qProducerBranchDelayBlockedBySuffixFiller;
-	unsigned int qProducerBranchDelaySuffixFillerTokenIndex;
+	bool qProducerBranchDelayBlockedBySuffixDependency;
+	unsigned int qProducerBranchDelaySuffixBlockerTokenIndex;
 	bool emitsDrain;
 	std::vector<unsigned int> prefetchTokenIndices;
 	std::vector<VuSoftwarePipelinePrefetch> prefetches;